fix(agent): preserve exhausted advice stream metadata
